made a separate TextComponent
